Add process and final methods to backends.
Since backend instances are stateful and can't be used to both encode
and decode at the same time, add process and final methods to them to
perform whatever task they're doing.  This simplifies the chain handling
code, since it no longer needs to compute or care what the method names
are.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>