vis: fix a decoding bug.
If a 0x9c byte appeared before another escape sequence, the regexp that
split apart the data to process in this chunk and the data to defer to
the next could split the second backslash in the "\M^\" sequence into a
separate chunk.  Adjust the regex slightly not to split apart a "^\" or
"\\" sequence, as this is probably going to do the wrong thing.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>