vis: clean up using switch
The previous code ended up being rather messy and hard to follow.  Clean
it up by using a switch statement.

This requires a dependency on the experimental module in order to handle
the feature properly, but this is a pure-Perl module that can easily be
fatpacked if necessary.  Using the poor man's no-warnings method makes
perlcritic desperately unhappy, so experimental it is.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>