Auto-generate LICENSE information
Dist::Zilla's License plugin is unhappy if you have a LICENSE file.
However, if you don't declare a license in the dist.ini file, it embeds
Perl 5 license information into the POD.  Fix this by using the License
plugin again (and therefore using the Basic plugin bundle) and adding
LICENSE to MANIFEST.SKIP.  This allows us to maintain the LICENSE file
in source so that it shows up on GitHub.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>