Change uri(form) to form
URI encoding tends to encode + as %2B, although there are places in URIs
where + is allowed (such as the path).  As a result, it isn't safe to
automatically decode + as space, as is done in
application/x-www-form-urlencoded, and the general rule is that any
variant of the encoding is decoded by the general variant of the
decoder.  Consequently, change the code that handles uri(form) and split
it out into form.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>