codec/registry: make registry no longer a singleton
It would be useful for people to be able to add additional types to a
codec registry, but making mutable singletons in Rust requires a lot of
ugliness which is hard to abstract over.  Instead, make it possible to
instantiate multiple instances of the registry to make it possible to
have them be easily mutable.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>