codec: handle truncated data
When we parse data, the codec may not read all of the input data we
provide. For example, we may not have sufficient space to write all of
the output data, or the input data may contain a partial sequence that
we cannot handle.

In such a situation, we must be careful to make forward progress and not
loop forever. If we get a BufError, signaling that we need additional
data, and we didn't consume any input bytes, check to see if we have the
same number of bytes in the input buffer as last time, and if so tell
the codec to flush itself. If we still make no progress, it means that
we have truncated data that cannot make a valid sequence, so report that
to the caller.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>