uri: fix off-by-one error
When we got a partial sequence, we subtracted one from the index of the
current input byte and returned that as the count of input bytes read.
However, that's incorrect. When we see a percent sign, the index of that
byte is the same as the number of bytes we've already consumed, because
the latter is a count, not an index. Correct this off-by-one error to
ensure we make progress.

In addition, add tests for this case and a few other invalid cases.