codec: adjust decoding pad computation
In the future, we're going to deal with formats that act like they are
padded, except without a real pad character. An example of such a format
is url64, the Base64 format with URL-safe encoding, which is typically
not padded. To prepare for this, let's adjust the way we compute padding
characters so that we consider input (i.e., for decoding) bytes instead
of pad characters. Rename the function which computes this value so it
more accurately reflects the value we're computing.

Since we're now performing complex operations on the PaddedEncoder and
PaddedDecoder, move the tests to the main codec tests module where they
belong. Provide comments about how we're splitting up this file now that
we have both tests and helper functions in the same module.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>