codec: make PaddedDecoder deal with odd-sized chunks
The PaddedDecoder class failed to deal with odd-sized chunks in several
places. First, it always returned a StreamEnd if we processed data,
which could cause the stream to be truncated. Second, if a small amount
of data was passed to the function, we could end up passing it through
to the underlying ChunkedDecoder, resulting in an overly long result.

Third, we would not handle a Finish flush state with no data and would
attempt to process no data, leading to a panic. And finally, we searched
the source buffer for a pad character farther than the amount of data we
processed, which could lead us to underflow in our subtraction.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>