codec: replace Buffer with simpler buffer solution
The Buffer class could end up copying a large amount of data due to the
way it was designed. This was obvious when we increased the buffer size
from 512 bytes to 8192 bytes and the uri encoder spent most of its time
in memcpy.

To deal with this problem, use a smaller, simpler solution: a buffer and
an offset where we read into the buffer until it's full, and then copy
the leftover bytes, if any, to the beginning of the buffer. This means
that we may end up doing a small amount of copying more frequently, but
we'll copy far less data overall, leading to far better performance.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>