codec: create a trait for codec implementations
In order to implement a useful help message, we'll need information from
each codec about what features it supports. In order to include this
with the factory function, implement a trait that transforms can
implement that provides the factory function and various other
information we can use in the help message.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>