codecs: add vis
Add the basic functionality of the vis(3) function for encoding and
decoding.  This implementation supports all features except
locale-sensitive comparisons, shell character recognition (although glob
character recognition is supported), no-slash support (which is not
invertible), safe character support, and URI and quoted-printable
encodings (which are or will be handled in their own codecs).

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>