chain: make Cursor use compatible with Rust 1.24
Rust 1.24 doesn't allow us to take a mutable reference to a Vec as a
data source, so switch to instantiating the Vec directly and then
calling into_inner on the Cursor at the end.  This is not only more
compatible, but it also is simpler and avoids needing an additional set
of indentation.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>