registry: allow iterating over codecs
In order to produce a list of codecs to list to the user, we need to
iterate over the existing codec registry.  Since we'll always want to
list the items in alphabetical order, use a BTreeMap instead of a
HashMap.  The performance penalty is not significant here, since we will
look up codecs only once in an average process.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>