codec: fix input truncation problem
The CodecReader implementation has a problem: if we read two full
buffers' worth of data, the last read size will equal to the current
read size, and we'll think we've hit end of file.  This causes data
larger than twice the buffer length to be unceremoniously truncated,
which we certainly don't want.  Fix this by removing this check and
instead relying on getting zero read bytes, which is reliable.

A test is not included in this commit, but the following commit will
introduce a new codec which implicitly tests this change.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>