codecs: introduce hash codec
Introduce a codec for performing hashes of a given piece of data.
Currently all the hashes are cryptographic in nature, but we may
introduce additional non-cryptographic hashes (e.g. CRC32) in a future
commit.  Add BLAKE2b and BLAKE2s in addition to the hashes supported in
the Perl version of muter.

Add a new error type that indicates a missing argument, since we don't
have a default hash type and therefore require an argument.

This commit implicitly tests the issue fixed in the previous commit by
processing data that is larger than the 8192-byte buffer used in the
CodecReader.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>