vis: use iterative decoding algorithm
When processing a series of encoded NUL bytes, it was possible in some
cases to overflow the stack due to a recursive call to handle_escape.
This could be triggered with a sequence of as few as 16384 NUL bytes.

Since stack overflows are undesirable, switch to always using an
iterative decoding algorithm.  Make our iterator peekable and only peek
ahead at the next byte so that we don't consume a following backslash
that we will want to operate on next round.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>