codec: add a trait method to convert Codecs to boxed BufReads
In every codec factory, we wrap the actual codec into a transform and
then box it.  This is lots of duplicated code, and there's never a
reason to do anything else.

Introduce a trait with a single method that turns the existing Codec
into a boxed BufRead.  Allow specifying the buffer size as well.  This
reduces the duplication, makes it more obvious what's going on, and
allows us to easily change the underlying implementation if we want to
in the future.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>