vis: improve handling of partial sequences
Due to the order of the matching for meta escape sequences, a couple of
bugs occurred when decoding.  First, partial escape sequences were
flagged as invalid sequences instead of truncated data.  Second, due to
the order of the patterns in the match block, if a "\M-x" or "\M^x"
sequence got cut off before the final character (due to block handling,
for example), an Err(Error::InvalidSequence) would be returned instead
of a Ok(Status::BufError), resulting in the data being flagged as
invalid.

Correct both of these by adjusting the match arms to include more cases
in a different order.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>