Add tests for invalid data
For encodings, all bytes are valid data, and handling this is
demonstrated by the round trip tests.  However, for decoders, we want to
be sure that we handle invalid data well.  In strict mode, we should
reject invalid input, and in non-strict mode, some invalid inputs may be
accepted.  Regardless, the one thing we want to be sure we don't do is
panic on invalid data.

Add some automated test cases that demonstrate invalid data and ensure
that whatever we do, we don't return an error.  Fix two separate cases
that did produce an error, including a place in the vis codec where we
didn't check whether certain bytes were octal digits and a place in the
uuencode code where we could subtract with overflow with invalid bytes.
We reject these bytes in strict mode, and allow them to be passed
through in non-strict mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>