Allow transform arguments to take values
In some cases, we want to allow users to adjust the length of the output
used by a transform.  To do so, allow arguments to be in the form
"key=value", where the key and value are both strings.  This allows us
the greatest flexibility to handle future situations without needing to
retrofit the existing code.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>