codecs: fix a data overwrite bug in PaddedEncoder
When in flush mode, the PaddedEncoder calls transform twice.  The first
time, it presumably writes some set of bytes into the output buffer.  If
we call it again, we need to account for that, and slice the output
buffer accordingly so we don't overwrite the data we got the first time.

This commit introduces a set of integration tests with multiple codecs
for improved testing.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>