codec: handle small chunks for PaddedEncoder
When we process a small chunk as the final chunk in the PaddedEncoder,
it's possible for the chunk to be exactly the input chunk size.  If so,
our attempt to process all but the last byte will return (0, 0), since
there isn't enough to handle a chunk.  In such a case, we need to
continue to the second transform call so that we end up processing any
remaining data.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>