codec: add checks for too-small output buffers
In the future, we'll let people control the buffer size as a
command-line option.  In order to make things work properly, we need to
verify that the output buffer will be large enough.  If it is not, we'll
get spurious errors, truncated data, or no output.

Add a method to each codec to specify the required output buffer size
and check it.  Add a special error type, BufError, which can be returned
by a transform method if for some reason the value must be detected at
that point.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>