Add a codec to append a newline
By default, muter does not append newlines to codec output because doing
so can cause ambiguous output.  However, many times a newline is
desired, such as when encoding Base64 data for transport to another
program.  Add a codec, lf, that simply appends a newline.  Add an
option, empty, that skips printing a newline if the output would be
empty, which is also typical behavior.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>