Makefile: allow unknown clippy lints
We will, in a future commit, disable a clippy lint that that exists only
in nightly.  However, Rust 1.47.0 will reject this lint since we reject
all warnings.  Let's tell clippy that we'd like to ignore unknown lints,
so that we can make our code work on both stable and nightly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>