codec: make options maps use a String
For a long time, we've specified the BTreeMaps that contain our options
to use a &'static str for the description of the argument.  However,
we're going to start translating those strings, and the result of the
translation is a String.  Let's change the type of our options so that
we can translate these strings in the future.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>