Include the tr crate for localization support
We want muter to be able to be localized because people should be able
to use a program in the language they prefer.  To make this easier,
let's include the tr crate, which calls out to an appropriate gettext
implementation.

To make things simpler, let's also use the system gettext if one is
available; if one is not, then the gettext-rs crate will ignore this
request and will build one for us.

Note that we don't localize any strings here.  Those will come in a
series of future commits as we go along.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>