Bump minimum Rust version to 1.34.2
Rust 1.24.1 is no longer supported by any of our supported Linux
distros.  All of them have moved on.  Let's move to 1.34.2, which is the
oldest version supported by any of our distros.

Note that 1.41.1 is our officially supported version, since it is the
version that is security-supported by our distros, but for now, let's
try to maintain a little more backwards compatibility if possible.

Since there aren't any distros that support this exact version, we'll
need to do a little more work.  We'll add a CI job to check the oldest
Rust version and build with that.

Finally, since we no longer need to maintain compatibility with older
versions of cargo, let's update the Cargo.lock to include the header
that newer versions include.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>