Makefile: build and test Debian packages during CI
Debian packages are a useful way to distribute software and it's
relatively easy to build them with cargo-deb.  Let's test that our
Debian packages remain buildable and that they're free of any Lintian
errors by running these jobs in CI, and let's add some Makefile targets
to make this easier.

Note that we now clean specific directories under target/ in the clippy
target because we cannot delete target/debian inside the Docker
container now that it's a volume mount.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>