doc/man: fix rendering issues
If we split the text of a link across lines, it won't be rendered
properly in the manual page due to a bug in Asciidoctor.  Let's make
sure our link text is all on one line.

While we're at it, let's strip out an extra space which is also unwanted
in our rendered output.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>