Cargo.toml: explicitly specify files to be included
This project uses AsciiDoc for documentation because it works well in a
variety of formats, as contrasted with Markdown, which is much more
closely tied with HTML.  However, for the README, crates.io wants it to
be in Markdown, so we build the Markdown version from the AsciiDoc
version.

This works fine, but cargo doesn't include it in our package, since it's
in .gitignore (as it should be).  So in order to include it, we'll need
to explicitly specify the files to be included instead of those to be
excluded.  Let's do so and tell cargo that we want to use README.md as
the readme for the archive, even if it is ignored.

Since this makes cargo upset that our README is not checked into Git,
let's pass the --allow-dirty flag.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>