codecs: add blake3 as hash option
BLAKE3 is a popular hash algorithm, and it is believed to be
cryptographically secure.  Include it as part of the hash codec, but
only when the non-default modern feature is enabled.  The blake3 crate
does not build with Rust 1.34.2 due to its dependency on a relatively
new version of arrayvec.

Deal with the fact that BLAKE3 is essentially an XOF by adjusting the
hash code to deal with that case.

In CI, build with modern features for stable, nightly, and macOS builds.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>