codecs/vis: detect incomplete octal sequences
vis(3) will always write an octal sequence with three octal characters
unless it is the special sequence "\0".  As a result, when decoding, we
can safely assume that all valid octal escape input will be one of those
two encodings.

However, we neglected to check that the last character in the escape was
actually a valid octal character.  Our invalid data test based on the
timestamp caught this as a subtract with overflow because in such a case
it was not: it was some non-octal character less than 0x30.

Let's check for this case and return an error.  We check for both the
case that the invalid character is larger bytewise than the intended
character range as well as smaller.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>