Makefile: add support for FreeBSD in CI
FreeBSD is a widely available platform and it's one of our supported
environments.  Let's add support for it in CI.

As part of this, we need to fix a bug in the Makefile.  When invoking
make, one must always use the MAKE variable and not the word "make"
directly in order to be sure one is invoking the correct program.  On
FreeBSD, "make" is the system make, which is a BSD-style make, whereas
we'd like GNU make for our builds.

We have to use a macOS runner here because Linux runners don't support
nested virtualization. Similarly, because we push the repository into
the VM with Git, be sure to clone the full repository on the system.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>