Makefile: add support for NetBSD in CI
NetBSD is a widely available platform and it's one of our supported
environments.  Let's add support for it in CI.

Note that because pkgsrc programs aren't installed under /usr/local, we
need to help cargo find the gettext location.  In addition,
Asciidoctor's binary name is suffixed with its Ruby version, so move
that into a variable as well.

We have to use a macOS runner here because Linux runners don't support
nested virtualization. Similarly, because we push the repository into
the VM with Git, be sure to clone the full repository on the system.

Disable HTTP multiplexing since it seems to be flaky on NetBSD's cargo.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>