Makefile: force writing gzipped data
By default, gzip complains if the destination file exists.  That means
if the Makefile rebuilds the documentation due to an out-of-date
dependency, that it will complain about attempting to overwrite the file
and prompt the user, which we don't want in a build system.

Let's fix this by forcing the overwrite of the files.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>