Makefile: fix CI on BSDs
In some cases, we can end up where we're not on a branch at all, since
GitHub Actions doesn't always end up with us on the branch we expected.
This can cause the push to fail because Git has no refs to push.

Let's fix this by always pushing HEAD to a fixed branch and then
checking that branch out inside our VM, which should end up with this
working properly whether we're on a branch or not.